refactor: 폴더 제거 #10
