--- src/components/common/Input/index.ts
@@ -1,1 +1,0 @@
-export { default as Input } from './Input';
